ui: publish dashboard page composition
