+++ .changeset/dashboard-page-composition.md
@@ -1,0 +1,5 @@
+---
+"@voyantjs/admin": minor
+---
+
+Publish the operator dashboard page composition, dashboard skeletons, and aggregate query helpers from `@voyantjs/admin`.
