+++ packages/admin/src/dashboard/dashboard-query-options.ts
@@ -1,0 +1,238 @@
+import { queryOptions } from "@tanstack/react-query"
+import type { VoyantFetcher } from "@voyantjs/react"
+import type { ChartConfig } from "@voyantjs/ui/components/chart"
+
+export interface DashboardQueryClient {
+  baseUrl: string
+  fetcher: VoyantFetcher
+}
+
+export type BookingsAggregates = {
+  total: number
+  totalPax: number
+  countsByStatus: Array<{ status: string; count: number }>
+  monthlyCounts: Array<{ yearMonth: string; count: number }>
+  monthlyRevenue: Array<{ yearMonth: string; currency: string; sellAmountCents: number }>
+  upcomingDepartures: {
+    count: number
+    items: Array<{
+      id: string
+      bookingNumber: string | null
+      status: string
+      startDate: string | null
+      endDate: string | null
+      pax: number | null
+      sellCurrency: string | null
+      sellAmountCents: number | null
+    }>
+  }
+}
+
+export type ProductsAggregates = {
+  total: number
+  active: number
+  publicActive: number
+  countsByStatus: Array<{ status: string; count: number }>
+  monthlyCreatedCounts: Array<{ yearMonth: string; count: number }>
+}
+
+export type SuppliersAggregates = {
+  total: number
+  active: number
+  countsByStatus: Array<{ status: string; count: number }>
+  countsByType: Array<{ type: string | null; count: number }>
+}
+
+export type FinanceAggregates = {
+  total: number
+  countsByStatus: Array<{ status: string; count: number }>
+  monthlyRevenue: Array<{ yearMonth: string; currency: string; totalCents: number }>
+  monthlyInvoiceCounts: Array<{ yearMonth: string; count: number }>
+  outstanding: Array<{ currency: string; balanceDueCents: number; count: number }>
+  overdue: Array<{ currency: string; balanceDueCents: number; count: number }>
+  outstandingTopN: Array<{
+    id: string
+    invoiceNumber: string | null
+    bookingId: string | null
+    status: string
+    currency: string
+    totalCents: number
+    balanceDueCents: number
+    issueDate: string | null
+    dueDate: string | null
+  }>
+}
+
+export class DashboardApiError extends Error {
+  constructor(
+    message: string,
+    public readonly status: number,
+    public readonly body: unknown,
+  ) {
+    super(message)
+    this.name = "DashboardApiError"
+  }
+}
+
+function buildSixMonthWindow() {
+  const now = new Date()
+  const fromDate = new Date(Date.UTC(now.getUTCFullYear(), now.getUTCMonth() - 5, 1, 0, 0, 0, 0))
+  return { from: fromDate.toISOString() }
+}
+
+function joinUrl(baseUrl: string, path: string): string {
+  const trimmedBase = baseUrl.endsWith("/") ? baseUrl.slice(0, -1) : baseUrl
+  const trimmedPath = path.startsWith("/") ? path : `/${path}`
+  return `${trimmedBase}${trimmedPath}`
+}
+
+function extractErrorMessage(status: number, statusText: string, body: unknown): string {
+  if (typeof body === "object" && body !== null && "error" in body) {
+    const errorField = (body as { error: unknown }).error
+    if (typeof errorField === "string") return errorField
+    if (typeof errorField === "object" && errorField !== null && "message" in errorField) {
+      return String((errorField as { message: unknown }).message)
+    }
+  }
+  return `Dashboard API error: ${status} ${statusText}`
+}
+
+async function readBody(response: Response): Promise<unknown> {
+  const text = await response.text()
+  if (!text) return undefined
+  try {
+    return JSON.parse(text)
+  } catch {
+    return text
+  }
+}
+
+async function fetchDashboardJson<T>(client: DashboardQueryClient, path: string): Promise<T> {
+  const response = await client.fetcher(joinUrl(client.baseUrl, path), {
+    headers: { Accept: "application/json" },
+    method: "GET",
+  })
+
+  const body = await readBody(response)
+
+  if (!response.ok) {
+    throw new DashboardApiError(
+      extractErrorMessage(response.status, response.statusText, body),
+      response.status,
+      body,
+    )
+  }
+
+  return body as T
+}
+
+export function getDashboardBookingsAggregatesQueryOptions(client: DashboardQueryClient) {
+  const { from } = buildSixMonthWindow()
+  return queryOptions({
+    queryKey: ["dashboard-bookings-aggregates", from],
+    queryFn: () =>
+      fetchDashboardJson<{ data: BookingsAggregates }>(
+        client,
+        `/v1/admin/bookings/aggregates?from=${encodeURIComponent(from)}&upcomingLimit=8`,
+      ),
+    staleTime: 60_000,
+  })
+}
+
+export function getDashboardProductsAggregatesQueryOptions(client: DashboardQueryClient) {
+  return queryOptions({
+    queryKey: ["dashboard-products-aggregates"],
+    queryFn: () =>
+      fetchDashboardJson<{ data: ProductsAggregates }>(client, "/v1/admin/products/aggregates"),
+    staleTime: 60_000,
+  })
+}
+
+export function getDashboardSuppliersAggregatesQueryOptions(client: DashboardQueryClient) {
+  return queryOptions({
+    queryKey: ["dashboard-suppliers-aggregates"],
+    queryFn: () =>
+      fetchDashboardJson<{ data: SuppliersAggregates }>(client, "/v1/admin/suppliers/aggregates"),
+    staleTime: 60_000,
+  })
+}
+
+export function getDashboardFinanceAggregatesQueryOptions(client: DashboardQueryClient) {
+  const { from } = buildSixMonthWindow()
+  return queryOptions({
+    queryKey: ["dashboard-finance-aggregates", from],
+    queryFn: () =>
+      fetchDashboardJson<{ data: FinanceAggregates }>(
+        client,
+        `/v1/admin/finance/aggregates?from=${encodeURIComponent(from)}&outstandingTopLimit=5`,
+      ),
+    staleTime: 60_000,
+  })
+}
+
+export function formatCurrency(cents: number, currency = "USD"): string {
+  return new Intl.NumberFormat("en-US", {
+    style: "currency",
+    currency,
+    maximumFractionDigits: 0,
+  }).format(cents / 100)
+}
+
+export function getStatusColor(status: string): string {
+  const map: Record<string, string> = {
+    confirmed: "var(--color-chart-1, hsl(142 71% 45%))",
+    completed: "var(--color-chart-2, hsl(221 83% 53%))",
+    in_progress: "var(--color-chart-3, hsl(47 96% 53%))",
+    draft: "var(--color-chart-4, hsl(215 14% 55%))",
+    cancelled: "var(--color-chart-5, hsl(0 84% 60%))",
+  }
+  return map[status] ?? "hsl(215 14% 55%)"
+}
+
+export const revenueChartConfig = {
+  revenue: { label: "Revenue", color: "hsl(221 83% 53%)" },
+  bookings: { label: "Bookings", color: "hsl(142 71% 45%)" },
+} satisfies ChartConfig
+
+export const bookingStatusConfig = {
+  confirmed: { label: "Confirmed", color: "hsl(142 71% 45%)" },
+  completed: { label: "Completed", color: "hsl(221 83% 53%)" },
+  in_progress: { label: "In Progress", color: "hsl(47 96% 53%)" },
+  draft: { label: "Draft", color: "hsl(215 14% 55%)" },
+  cancelled: { label: "Cancelled", color: "hsl(0 84% 60%)" },
+} satisfies ChartConfig
+
+export const monthlyBookingsConfig = {
+  count: { label: "Bookings", color: "hsl(221 83% 53%)" },
+} satisfies ChartConfig
+
+export function buildMonthSeries() {
+  const now = new Date()
+  return Array.from({ length: 6 }, (_, idx) => {
+    const offset = 5 - idx
+    const date = new Date(Date.UTC(now.getUTCFullYear(), now.getUTCMonth() - offset, 1))
+    return {
+      yearMonth: `${date.getUTCFullYear()}-${String(date.getUTCMonth() + 1).padStart(2, "0")}`,
+      month: date.toLocaleString("en-US", { month: "short", timeZone: "UTC" }),
+    }
+  })
+}
+
+export function pickPrimaryCurrency(
+  rows: Array<{ currency: string; sellAmountCents: number }>,
+): string {
+  if (rows.length === 0) return "USD"
+  const totals = new Map<string, number>()
+  for (const row of rows) {
+    totals.set(row.currency, (totals.get(row.currency) ?? 0) + row.sellAmountCents)
+  }
+  let bestCurrency = "USD"
+  let bestTotal = -1
+  for (const [currency, total] of totals) {
+    if (total > bestTotal) {
+      bestCurrency = currency
+      bestTotal = total
+    }
+  }
+  return bestCurrency
+}
