+++ templates/operator/src/routes/_workspace/index.tsx
@@ -1,20 +1,32 @@
 import { createFileRoute } from "@tanstack/react-router"
-import { DashboardPage } from "@/components/voyant/dashboard/dashboard-page"
 import {
+  DashboardPage,
+  DashboardSkeleton,
   getDashboardBookingsAggregatesQueryOptions,
   getDashboardFinanceAggregatesQueryOptions,
   getDashboardProductsAggregatesQueryOptions,
   getDashboardSuppliersAggregatesQueryOptions,
-} from "@/components/voyant/dashboard/dashboard-shared"
-import { DashboardSkeleton } from "@/components/voyant/dashboard/dashboard-skeleton"
+} from "@voyantjs/admin"
+import { defaultFetcher } from "@voyantjs/react"
+import { getApiUrl } from "@/lib/env"
+
+const dashboardClient = { baseUrl: getApiUrl(), fetcher: defaultFetcher }
 
 export const Route = createFileRoute("/_workspace/")({
   loader: ({ context }) =>
     Promise.all([
-      context.queryClient.ensureQueryData(getDashboardBookingsAggregatesQueryOptions()),
-      context.queryClient.ensureQueryData(getDashboardProductsAggregatesQueryOptions()),
-      context.queryClient.ensureQueryData(getDashboardSuppliersAggregatesQueryOptions()),
-      context.queryClient.ensureQueryData(getDashboardFinanceAggregatesQueryOptions()),
+      context.queryClient.ensureQueryData(
+        getDashboardBookingsAggregatesQueryOptions(dashboardClient),
+      ),
+      context.queryClient.ensureQueryData(
+        getDashboardProductsAggregatesQueryOptions(dashboardClient),
+      ),
+      context.queryClient.ensureQueryData(
+        getDashboardSuppliersAggregatesQueryOptions(dashboardClient),
+      ),
+      context.queryClient.ensureQueryData(
+        getDashboardFinanceAggregatesQueryOptions(dashboardClient),
+      ),
     ]),
   pendingComponent: DashboardSkeleton,
   component: DashboardPage,
